+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+MINOR
+- Add Element .parent, .tagName; add CssDeclaration .overflowX/Y
+
 # Changelog
 ## 1.21.0
 - Add AnchorElement and Element.click()